Report details of errors interacting with GitHub API
Specifically motivated by the need to get to the bottom of a Gist export
problem that a user is experiencing. In general, if we get an error from
the API, we’ll send a report to Bugsnag that includes details of the
failed request in the metadata.

Ideally, we’d also handle the error gracefully in the UI (show an error
message, don’t open blank browser tab), but at this point I just want to
get to the bottom of the underlying error in question.