Mildly more verbose webpack output
Webpack was reporting a “compiled with warnings” but not showing any
actual warnings, because the output was set to `errors-only` which is
very terse. Configure the output for both dev and test to be a bit more
verbose, in particular including warnings, traces for both warning and
errors, and top-line information about the chunks built etc.